Consolidate Kafka packages into one single package
These 2 packages `pkg/reconciler/kafka` and `pkg/kafka` always
confuse me about their purpose, so I think it makes more sense to
keep one single package for "Kafka" related functions.
