+++ control-plane/pkg/kafka/consumer_group.go
