+++ control-plane/pkg/security/secret.go
@@ -23,7 +23,8 @@ import (
 	"strconv"
 
 	"github.com/Shopify/sarama"
-	"knative.dev/eventing-kafka-broker/control-plane/pkg/reconciler/kafka"
+
+	"knative.dev/eventing-kafka-broker/control-plane/pkg/kafka"
 )
 
 const (
