+++ test/test_images/kafka-consumer/consumer.go
@@ -33,7 +33,7 @@ import (
 	"knative.dev/pkg/signals"
 
 	eventing "knative.dev/eventing-kafka-broker/control-plane/pkg/apis/eventing/v1alpha1"
-	"knative.dev/eventing-kafka-broker/control-plane/pkg/reconciler/kafka"
+	"knative.dev/eventing-kafka-broker/control-plane/pkg/kafka"
 	kafkatest "knative.dev/eventing-kafka-broker/test/pkg/kafka"
 )
