+++ test/upgrade/continual/broker.go
@@ -35,7 +35,7 @@ import (
 
 	eventingkafkaupgrade "knative.dev/eventing-kafka/test/upgrade/continual"
 
-	"knative.dev/eventing-kafka-broker/control-plane/pkg/reconciler/kafka"
+	"knative.dev/eventing-kafka-broker/control-plane/pkg/kafka"
 	testingpkg "knative.dev/eventing-kafka-broker/test/pkg/testing"
 )
